updated 3/27

+++ mar_scripts/manet/mmaction2/mmaction/datasets/pipelines/loading.py
@@ -252,7 +252,7 @@ class SampleFrames:
 
         start_index = results['start_index']
         frame_inds = np.concatenate(frame_inds) + start_index
-        results['frame_inds'] = frame_inds.astype(np.int)
+        results['frame_inds'] = frame_inds.astype(np.int32)
         results['clip_len'] = self.clip_len
         results['frame_interval'] = self.frame_interval
         results['num_clips'] = self.num_clips
